cleaned up based on PR comments
